# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: api-lifecare-integrator
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
- url: http://localhost:61256
  description: Generated server url
...
paths:
...
  /{municipalityId}/actualisations:
    get:
      tags:
      - Actualisations
      description: Get the actualisations (case intakes) registered on a person in
        the given period
      operationId: getActualisations
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: partyId
        in: query
        description: Party id of the person
        required: true
        schema:
          type: string
        example: 81471222-5798-11e9-ae24-57fa13b361e1
      - name: from
        in: query
        description: Start of the period
        required: false
        schema:
          type: string
          format: date
        example: 2025-01-01
      - name: to
        in: query
        description: End of the period
        required: false
        schema:
          type: string
          format: date
        example: 2026-12-31
      - name: page
        in: query
        description: Page number
        required: false
        schema:
          type: integer
          format: int32
          exclusiveMinimum: 0
        example: 1
      - name: pageSize
        in: query
        description: Page size
        required: false
        schema:
          type: integer
          format: int32
          exclusiveMinimum: 0
          maximum: 1000
        example: 20
      - name: ascending
        in: query
        description: Sort order
        required: false
        schema:
          type: boolean
        example: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponseActualisation"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      tags:
      - Actualisations
      description: Create an actualisation (case intake) in Lifecare family care.
        Ids in the request come from the proposal endpoint
      operationId: createActualisation
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActualisationRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedResource"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/actualisations/{actualisationId}/attachments:
    post:
      tags:
      - Actualisations
      description: Upload a document and bind it to an actualisation. Type codes come
        from the proposal endpoint's attachmentTypes
      operationId: addActualisationAttachment
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: actualisationId
        in: path
        description: Id of the actualisation
        required: true
        schema:
          type: integer
          format: int32
        example: 12345
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                documentType:
                  type: string
                  description: The document type code
                  minLength: 1
                senderType:
                  type: string
                  description: The document sender type code
                  minLength: 1
                title:
                  type: string
                  description: The document title
                senderName:
                  type: string
                  description: The sender name
                file:
                  type: string
                  format: binary
              required:
              - documentType
              - file
              - senderType
        required: true
      responses:
        "204":
          description: No content
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/actualisations/proposal:
    get:
      tags:
      - Actualisations
      description: Get the proposal (valid code lists) needed to create an actualisation
        for a person
      operationId: getActualisationProposal
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: partyId
        in: query
        description: Party id of the person
        required: true
        schema:
          type: string
        example: 81471222-5798-11e9-ae24-57fa13b361e1
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActualisationProposal"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
...
    ConstraintViolationProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        causeAsProblem:
          $ref: "#/components/schemas/ThrowableProblem"
        detail:
          type: string
        instance:
          type: string
          format: uri
    ThrowableProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        causeAsProblem: {}
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    CreatedResource:
      type: object
      description: The id of the created resource in Lifecare FC
      properties:
        id:
          type: integer
          format: int32
          description: The id of the created resource in Lifecare FC
          examples:
          - 12345
    CreateActualisationRequest:
      type: object
      description: Request to create an actualisation in Lifecare FC. All ids come
        from the actualisation proposal endpoint.
      properties:
        partyId:
          type: string
          description: The party id of the person the actualisation concerns
          examples:
          - 81471222-5798-11e9-ae24-57fa13b361e1
        date:
          type: string
          format: date
          description: The date of the actualisation
          examples:
          - 2026-05-01
        typeId:
          type: integer
          format: int32
          description: "The id of the actualisation type, from the actualisation proposal\
            \ endpoint"
          examples:
          - 1
        fromWhoId:
          type: integer
          format: int32
          description: "The id of the from-who value, from the actualisation proposal\
            \ endpoint"
          examples:
          - 1
        reasonId:
          type: integer
          format: int32
          description: "The id of the reason, from the actualisation proposal endpoint"
          examples:
          - 1
        organisationId:
          type: integer
          format: int32
          description: "The id of the organization, from the actualisation proposal\
            \ endpoint"
          examples:
          - 1
        organisationUnitId:
          type: string
          description: "The id of the organization unit, from the actualisation proposal\
            \ endpoint"
          examples:
          - "100"
        caseworkerId:
          type: string
          description: "The id of the caseworker, from the actualisation proposal\
            \ endpoint"
          examples:
          - abc123
        specifiesId:
          type: integer
          format: int32
          description: "The id of the specify type, from the actualisation proposal\
            \ endpoint"
          examples:
          - 1
        serviceId:
          type: integer
          format: int32
          description: "The id of the service to link to, from the actualisation proposal\
            \ endpoint"
          examples:
          - 12345
        investigationId:
          type: integer
          format: int32
          description: "The id of the investigation to link to, from the actualisation\
            \ proposal endpoint"
          examples:
          - 12345
        workingStatusId:
          type: integer
          format: int32
          description: "The id of the working status, from the actualisation proposal\
            \ endpoint"
          examples:
          - 1
      required:
      - date
      - partyId
      - typeId
...
    Lookup:
      type: object
      description: A lookup value from a code list
      properties:
        id:
          type: integer
          format: int32
          description: The id of the lookup value
          examples:
          - 1
        name:
          type: string
          description: The display name of the lookup value
          examples:
          - Ansökan
...
    ProposalCase:
      type: object
      description: An investigation or service that can be linked to in a proposal
      properties:
        id:
          type: integer
          format: int32
          description: The id of the investigation or service
          examples:
          - 12345
        type:
          type: integer
          format: int32
          description: The type of the investigation or service
          examples:
          - 1
        name:
          type: string
          description: The name of the investigation or service
          examples:
          - Utredning ekonomiskt bistånd
        startDate:
          type: string
          format: date
          description: The start date of the investigation or service
          examples:
          - 2026-05-01
        organisationId:
          type: integer
          format: int32
          description: The id of the organization the investigation or service belongs
            to
          examples:
          - 1
        organisationUnitId:
          type: string
          description: The id of the organization unit the investigation or service
            belongs to
          examples:
          - "100"
        caseworkerId:
          type: string
          description: The id of the caseworker responsible for the investigation
            or service
          examples:
          - abc123
    Actualisation:
      type: object
      description: "An actualisation (aktualisering) from Lifecare Family Care, re-modelled\
        \ to a common shape"
      properties:
        id:
          type: integer
          format: int32
          description: The actualisation id in the source system
          examples:
          - 12345
        type:
          type: string
          description: The actualisation type
          examples:
          - Ansökan
        name:
          type: string
          description: Name of the person the actualisation concerns
          examples:
          - Kalle Karlsson
        date:
          type: string
          format: date
          description: The date of the actualisation
          examples:
          - 2026-05-01
        reason:
          type: string
          description: The reason for the actualisation
          examples:
          - Ansökan om ekonomiskt bistånd
        regards:
          type: string
          description: What the actualisation regards
          examples:
          - Ekonomiskt bistånd
        fromWho:
          type: string
          description: Who initiated the actualisation
          examples:
          - Den enskilde
        caseworker:
          type: string
          description: Name of the caseworker handling the actualisation
          examples:
          - Anna Andersson
        organization:
          type: string
          description: The organization handling the actualisation
          examples:
          - Vuxen försörjningsstöd
        status:
          type: string
          description: The actualisation status
          examples:
          - Inledd utredning
        investigationId:
          type: integer
          format: int32
          description: The id of the investigation connected to the actualisation
          examples:
          - 456
        serviceId:
          type: integer
          format: int32
          description: The id of the service connected to the actualisation
          examples:
          - 789
        decisionId:
          type: integer
          format: int32
          description: The id of the decision connected to the actualisation
          examples:
          - 1011
    PagedResponseActualisation:
      type: object
      description: "Paged response wrapper, mirroring the pagination of the underlying\
        \ Lifecare family care API"
      properties:
        page:
          type: integer
          format: int32
          description: The current page (one-based)
          examples:
          - 1
        pageSize:
          type: integer
          format: int32
          description: The number of records per page
          examples:
          - 20
        totalPages:
          type: integer
          format: int32
          description: The total number of pages
          examples:
          - 5
        totalRecords:
          type: integer
          format: int32
          description: The total number of records
          examples:
          - 98
        results:
          type: array
          description: The records on the current page
          items:
            $ref: "#/components/schemas/Actualisation"
    ActualisationProposal:
      type: object
      description: A proposal with the valid values needed to create an actualisation
        in Lifecare FC
      properties:
        actualisationTypes:
          type: array
          description: The valid actualisation types
          items:
            $ref: "#/components/schemas/ActualisationType"
        specifyTypes:
          type: array
          description: The valid specify types
          items:
            $ref: "#/components/schemas/Lookup"
        workingStatus:
          type: array
          description: The valid working statuses
          items:
            $ref: "#/components/schemas/Lookup"
        organizations:
          type: array
          description: The valid organizations
          items:
            $ref: "#/components/schemas/Organization"
        investigations:
          type: array
          description: The investigations that can be linked to
          items:
            $ref: "#/components/schemas/ProposalCase"
        services:
          type: array
          description: The services that can be linked to
          items:
            $ref: "#/components/schemas/ProposalCase"
        attachmentTypes:
          type: array
          description: The valid attachment types
          items:
            $ref: "#/components/schemas/AttachmentType"
    ActualisationType:
      type: object
      description: An actualisation type with its valid sub-code-lists
      properties:
        id:
          type: integer
          format: int32
          description: The id of the actualisation type
          examples:
          - 1
        name:
          type: string
          description: The name of the actualisation type
          examples:
          - Ansökan
        specifyTypeMandatory:
          type: boolean
          description: Whether a specify type must be given for this actualisation
            type
          examples:
          - "true"
        workingStatus:
          type: boolean
          description: Whether a working status must be given for this actualisation
            type
          examples:
          - "false"
        reasons:
          type: array
          description: The valid reasons for this actualisation type
          items:
            $ref: "#/components/schemas/Lookup"
        fromWho:
          type: array
          description: The valid from-who values for this actualisation type
          items:
            $ref: "#/components/schemas/Lookup"
        investigationTypes:
          type: array
          description: The valid investigation types for this actualisation type
          items:
            $ref: "#/components/schemas/Lookup"
        serviceTypes:
          type: array
          description: The valid service types for this actualisation type
          items:
            $ref: "#/components/schemas/Lookup"
    AttachmentType:
      type: object
      description: An attachment type with its valid sender types
      properties:
        id:
          type: integer
          format: int32
          description: The id of the attachment type
          examples:
          - 1
        name:
          type: string
          description: The name of the attachment type
          examples:
          - Ansökan
        senderTypes:
          type: array
          description: The valid sender types for this attachment type
          items:
            $ref: "#/components/schemas/Lookup"
    Organization:
      type: object
      description: An organization entry
      properties:
        id:
          type: integer
          format: int32
          description: The id of the organization
          examples:
          - 1
        unitId:
          type: string
          description: The id of the organization unit
          examples:
          - "100"
        name:
          type: string
          description: The name of the organization
          examples:
          - Vuxenenheten
  securitySchemes: {}